模块化重构，已测试未见异常
