--- models.py
@@ -1,248 +1,0 @@
-"""Data models and constants for ChuanHuaTong plugin."""
-
-from dataclasses import dataclass
-from typing import Any, Dict
-
-import astrbot.api.message_components as Comp
-
-PLAIN_COMPONENT_TYPES = tuple(
-    getattr(Comp, name)
-    for name in ("Plain", "Text")
-    if hasattr(Comp, name)
-)
-LINEBREAK_COMPONENT = getattr(Comp, "LineBreak", None)
-IMAGE_SUFFIXES = {".png", ".jpg", ".jpeg", ".webp"}
-
-
-@dataclass
-class EmotionMeta:
-    key: str
-    folder: str
-    label: str
-    color: str
-    enabled: bool = True
-
-
-DEFAULT_EMOTIONS: list[dict[str, Any]] = [
-    {"key": "neutral", "folder": "shy", "label": "平静", "color": "#A9C5FF", "enabled": True},
-    {"key": "happy", "folder": "happy", "label": "开心", "color": "#FFC857", "enabled": True},
-    {"key": "sad", "folder": "sad", "label": "低落", "color": "#7DA1FF", "enabled": True},
-    {"key": "shy", "folder": "shy", "label": "害羞", "color": "#F9C5D1", "enabled": True},
-    {"key": "surprise", "folder": "surprise", "label": "惊讶", "color": "#F5E960", "enabled": True},
-    {"key": "angry", "folder": "sad", "label": "生气", "color": "#FF8A8A", "enabled": True},
-]
-
-DEFAULT_PROMPT_TEMPLATE = (
-    "请在回答正文中就近插入一个情绪标签，例如 {tags}。"
-    "标签写在对应句子旁即可，便于渲染立绘。标签仅包含字母、数字或下划线。"
-)
-
-DEFAULT_LAYOUT: Dict[str, Any] = {
-    "canvas_width": 1600,
-    "canvas_height": 600,
-    "background_color": "#05060a",
-    "background_asset": "__auto__",
-    "box_left": 372.987132257021,
-    "box_top": 284.43201385917496,
-    "box_width": 1227.1886120996442,
-    "box_height": 317.58007117437717,
-    "padding": 28,
-    "font_size": 56,
-    "line_height": 1.1,
-    "align": "left",
-    "radius": 26,
-    "text_color": "#ffffff",
-    "text_stroke_width": 1,
-    "text_stroke_color": "#ffffff",
-    "text_bg": "rgba(0,0,0,0.52)",
-    "text_shadow": "0 3px 12px rgba(0,0,0,0.55)",
-    "body_font": "方正风雅宋简体.ttf",
-    "textbox_z_index": 130,
-    "character_asset": "__auto__",
-    "character_left": 1.0676156583630245,
-    "character_bottom": 0,
-    "character_top": 0,
-    "character_width": 499.72719967439286,
-    "character_z_index": 140,
-    "character_role": "__auto__",
-    "character_fit_mode": "fixed_width",
-    "character_uniform_height": 620,
-    "character_align_bottom": True,
-    "character_shadow": "drop-shadow(0 12px 36px rgba(0,0,0,0.6))",
-    "background_group": "__auto__",
-    "box_anchor": "left-top",
-    "character_anchor": "left-top",
-    "text_overlays": [
-        {
-            "id": "ov_1764055391684",
-            "type": "glass",
-            "text": "",
-            "image": "",
-            "font": "",
-            "left": 283,
-            "top": 282,
-            "width": 1317,
-            "height": 354,
-            "font_size": 28,
-            "color": "#ffffff",
-            "stroke_width": 0,
-            "stroke_color": "#000000",
-            "bold": True,
-            "z_index": 100,
-            "visible": True,
-            "opacity": 1.0,
-        },
-        {
-            "id": "ov_1764055442570",
-            "type": "image",
-            "text": "",
-            "image": "名称框.png",
-            "font": "",
-            "left": 410,
-            "top": 166,
-            "width": 150,
-            "height": 164,
-            "font_size": 28,
-            "color": "#FFFFFF",
-            "stroke_width": 0,
-            "stroke_color": "#000000",
-            "bold": True,
-            "z_index": 110,
-            "visible": True,
-            "opacity": 1.0,
-        },
-        {
-            "id": "ov_1764055449125",
-            "type": "image",
-            "text": "",
-            "image": "底框.png",
-            "font": "",
-            "left": -18,
-            "top": 169,
-            "width": 1632,
-            "height": 495,
-            "font_size": 24,
-            "color": "#ffffff",
-            "stroke_width": 0,
-            "stroke_color": "#000000",
-            "bold": True,
-            "z_index": 120,
-            "visible": True,
-            "opacity": 1.0,
-        },
-        {
-            "id": "ov_1764055563203",
-            "type": "text",
-            "text": "樱",
-            "image": "",
-            "font": "方正风雅宋简体.ttf",
-            "left": 450,
-            "top": 201,
-            "width": 90,
-            "height": 72,
-            "font_size": 72,
-            "color": "#b66363",
-            "stroke_width": 1,
-            "stroke_color": "#763737",
-            "bold": True,
-            "z_index": 180,
-            "visible": True,
-            "opacity": 1.0,
-        },
-        {
-            "id": "ov_1764055643818",
-            "type": "text",
-            "text": "羽",
-            "image": "",
-            "font": "方正风雅宋简体.ttf",
-            "left": 525,
-            "top": 222,
-            "width": 56,
-            "height": 41,
-            "font_size": 50,
-            "color": "#ffffff",
-            "stroke_width": 1,
-            "stroke_color": "#d6d6d6",
-            "bold": True,
-            "z_index": 170,
-            "visible": True,
-            "opacity": 1.0,
-        },
-        {
-            "id": "ov_1764055668857",
-            "type": "text",
-            "text": "艾",
-            "image": "",
-            "font": "方正风雅宋简体.ttf",
-            "left": 579,
-            "top": 206,
-            "width": 69,
-            "height": 68,
-            "font_size": 64,
-            "color": "#ffffff",
-            "stroke_width": 1,
-            "stroke_color": "#d9d9d9",
-            "bold": True,
-            "z_index": 160,
-            "visible": True,
-            "opacity": 1.0,
-        },
-        {
-            "id": "ov_1764055669272",
-            "type": "text",
-            "text": "玛",
-            "image": "",
-            "font": "方正风雅宋简体.ttf",
-            "left": 652,
-            "top": 221,
-            "width": 50,
-            "height": 48,
-            "font_size": 50,
-            "color": "#ffffff",
-            "stroke_width": 1,
-            "stroke_color": "#c7c7c7",
-            "bold": True,
-            "z_index": 150,
-            "visible": True,
-            "opacity": 1.0,
-        },
-        {
-            "id": "ov_1764057760732",
-            "type": "image",
-            "text": "",
-            "image": "线索.png",
-            "font": "",
-            "left": 1382,
-            "top": 0,
-            "width": 104,
-            "height": 105,
-            "font_size": 24,
-            "color": "#ffffff",
-            "stroke_width": 0,
-            "stroke_color": "#000000",
-            "bold": True,
-            "z_index": 200,
-            "visible": True,
-            "opacity": 1.0,
-        },
-        {
-            "id": "ov_1764057771825",
-            "type": "image",
-            "text": "",
-            "image": "设置.png",
-            "font": "",
-            "left": 1486,
-            "top": 3,
-            "width": 100,
-            "height": 100,
-            "font_size": 24,
-            "color": "#ffffff",
-            "stroke_width": 0,
-            "stroke_color": "#000000",
-            "bold": True,
-            "z_index": 190,
-            "visible": True,
-            "opacity": 1.0,
-        },
-    ],
-}
